Strips out personas
