+++ components/Contexts.component.tsx
@@ -7,10 +7,6 @@ import { Authentication, useAuthContext, AuthContextType } from './authenticate'
 import { ChatContextType } from '@/components/Chat/context'
 
 
-import { PersonaContextType, PersonaContext, PersonaPanel, usePersonaContext, PersonaModal  } from '@/components/Persona'
-
-import { AdminContextType, AdminContext, useAdminContext, AdminPanel } from '@/components/Admin'
-
 interface Props {
   children: ReactNode;
 }
@@ -18,17 +14,11 @@ interface Props {
 const Contexts = ({children}:Props)=>{
   const chatProvider: ChatContextType = useChatContext()
   const authProvider: AuthContextType = useAuthContext()
-  const personaProvider: PersonaContextType = usePersonaContext()
-  const adminProvider: AdminContextType = useAdminContext()
 
   return (
     <Authentication.Provider value={authProvider}>
       <ChatContext.Provider value={chatProvider}>
-        <PersonaContext.Provider value={personaProvider}>
-          <AdminContext.Provider value={adminProvider}>
-            {children}
-          </AdminContext.Provider>
-        </PersonaContext.Provider>
+\         {children}
       </ChatContext.Provider>
     </Authentication.Provider>
   )
