--- components/Persona/index.ts
@@ -1,2 +1,0 @@
-export { default as PersonaPanel } from './PersonaPanel.component'
-export { PersonaContext, usePersonaContext } from './persona.context'
