+++ components/index.ts
@@ -4,6 +4,5 @@ export * from './MobileMenu'
 export * from './Spin'
 export * from './Themes'
 export * from './Toaster'
-export * from './Persona'
 export * from './Banner'
 export * from './Sidebar'
